implemented time function fully, removed Passera's UInt, and modified Ordering for easier queue removal
Passera's Unsigned package is deprecated, facilitating the need to create a new package from scratch

time.scala is no longer a separate file

modified Ordering so front of queue will display the lowest usageFrequency, making the dequeue in case of reaching max topics easier. Updated Todos to reflect need for a priority double ended queue
